@timothycarambat 2024-09-24
@blazeyo Thanks a ton for handling this. Expanded scope to do the same functionality with reporting of `watched` status on documents, which would also cause issues if thousands were watched. While I'm sure we can improve the speed even further from 2sec - this is still incredible.

Again, thanks a ton.